+++ platform/jvm/capture-plugin/src/test/kotlin/io/bitdrift/capture/instrumentation/fakes/TestSpanAddingParameters.kt
@@ -34,7 +34,7 @@
 package io.bitdrift.capture.instrumentation.fakes
 
 import io.bitdrift.capture.extension.InstrumentationExtension.OkHttpInstrumentationType
-import io.bitdrift.capture.extension.InstrumentationExtension.WebViewAutomaticInstrumentationMode
+import io.bitdrift.capture.extension.InstrumentationExtension.WebViewAutomaticInstrumentationScope
 import io.bitdrift.capture.instrumentation.ClassInstrumentable
 import io.bitdrift.capture.instrumentation.SpanAddingClassVisitorFactory
 import org.gradle.api.internal.provider.DefaultProperty
@@ -45,7 +45,8 @@ import java.io.File
 class TestSpanAddingParameters(
     private val debugOutput: Boolean = true,
     private val okHttpInstrumentationEnabled: Boolean = true,
-    private val webViewInstrumentationModeValue: WebViewAutomaticInstrumentationMode? = WebViewAutomaticInstrumentationMode.FULL,
+    private val webViewInstrumentationScopeValue: WebViewAutomaticInstrumentationScope? =
+        WebViewAutomaticInstrumentationScope.ALL,
     private val inMemoryDir: File,
 ) : SpanAddingClassVisitorFactory.SpanAddingParameters {
     override val debug: Property<Boolean>
@@ -62,10 +63,10 @@ class TestSpanAddingParameters(
             DefaultProperty(PropertyHost.NO_OP, Boolean::class.javaObjectType)
                 .convention(okHttpInstrumentationEnabled)
 
-    override val webViewInstrumentationMode: Property<WebViewAutomaticInstrumentationMode>
+    override val webViewInstrumentationScope: Property<WebViewAutomaticInstrumentationScope>
         get() =
-            DefaultProperty(PropertyHost.NO_OP, WebViewAutomaticInstrumentationMode::class.java)
-                .also { property -> webViewInstrumentationModeValue?.let(property::set) }
+            DefaultProperty(PropertyHost.NO_OP, WebViewAutomaticInstrumentationScope::class.java)
+                .also { property -> webViewInstrumentationScopeValue?.let(property::set) }
 
     override val tmpDir: Property<File>
         get() = DefaultProperty<File>(PropertyHost.NO_OP, File::class.java).convention(inMemoryDir)